backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend for Skyent-dev"
authors = ["Xenolab <contact@xenolab.co>"]
readme = "README.md"
license = "MIT"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
pydantic-settings = "^2.2.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.28"}
alembic = "^1.13.1"
psycopg = {extras = ["binary"], version = "^3.1.18"}
asyncpg = "^0.29.0"
greenlet = "^3.0.3"
aiosqlite = "^0.19.0"
pandas = "^2.2.3"
spacy = ">=3.8.0,<3.9.0"
sumy = "^0.11.0"
tavily-python = "^0.7.2"
openai = "^1.78.1"
anthropic = "^0.51.0"
detoxify = "^0.5.2"
python-dotenv = "^1.1.0"
langsmith = "^0.3.42"
linkedin-api = "^2.3.1"
tweepy = "^4.15.0"
facebook-sdk = "^3.1.0"
google-api-python-client = "^2.169.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
ruff = "^0.3.0"
mypy = "^1.9.0"
httpx = "^0.27.0" # For TestClient
pytest-asyncio = "^0.23.5" # Pour les tests asynchrones

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # Pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = ["E501"] # line too long, handled by black

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
strict = true
ignore_missing_imports = true
